+++ src/View/Helper/LibraryHelper.php
@@ -12,6 +12,7 @@
  */
 namespace MeTools\View\Helper;
 
+use Cake\Core\Plugin;
 use Cake\Event\Event;
 use Cake\I18n\I18n;
 use Cake\View\Helper;
@@ -22,12 +23,14 @@ use Cake\View\Helper;
 class LibraryHelper extends Helper
 {
     /**
-     * Helpers
+     * Helpers.
+     *
+     * The `Asset` helper will be loaded by the `initialize()` method. If the
+     *  `Assets` plugin doesn't exist, it will be a copy of the `Html` helper.
      * @var array
      */
     public $helpers = [
-        ASSETS . '.Asset',
-        'Html' => ['className' => ME_TOOLS . '.Html'],
+        'Html' => ['className' => 'MeTools.Html'],
     ];
 
     /**
@@ -35,6 +38,26 @@ class LibraryHelper extends Helper
      * @var array
      */
     protected $output = [];
+
+    /**
+     * Constructor hook method.
+     *
+     * Implement this method to avoid having to overwrite the constructor and
+     *  call parent.
+     * @param array $config The configuration settings provided to this helper
+     * @return void
+     * @since 2.18.0
+     */
+    public function initialize(array $config)
+    {
+        parent::initialize($config);
+
+        if (Plugin::getCollection()->has('Assets')) {
+            $this->Asset = $this->getView()->loadHelper('Assets.Asset');
+        } else {
+            $this->Asset = clone $this->Html;
+        }
+    }
 
     /**
      * Internal function to generate datepicker and timepicker.
@@ -51,7 +74,7 @@ class LibraryHelper extends Helper
     {
         $this->Asset->script([
             '/vendor/moment/moment-with-locales.min',
-            ME_TOOLS . '.bootstrap-datetimepicker.min',
+            'MeTools.bootstrap-datetimepicker.min',
         ], ['block' => 'script_bottom']);
 
         $this->Asset->css(
@@ -116,7 +139,7 @@ class LibraryHelper extends Helper
      */
     public function analytics($id)
     {
-        return $this->getView()->request->is('localhost') ? null : $this->Html->scriptBlock(
+        return $this->getView()->getRequest()->is('localhost') ? null : $this->Html->scriptBlock(
             sprintf('!function(e,a,t,n,c,o,s){e.GoogleAnalyticsObject=c,e[c]=e[c]||function(){(e[c].q=e[c].q||[]).push(arguments)},e[c].l=1*new Date,o=a.createElement(t),s=a.getElementsByTagName(t)[0],o.async=1,o.src=n,s.parentNode.insertBefore(o,s)}(window,document,"script","//www.google-analytics.com/analytics.js","ga"),ga("create","%s","auto"),ga("send","pageview");', $id),
             ['block' => 'script_bottom']
         );