+++ tests/TestCase/TestSuite/MockTraitTest.php
@@ -10,19 +10,17 @@
  * @link        https://github.com/mirko-pagliai/me-tools
  * @license     https://opensource.org/licenses/mit-license.php MIT License
  */
-namespace MeTools\Test\TestCase\TestSuite\Traits;
+namespace MeTools\Test\TestCase\TestSuite;
 
 use Cake\ORM\TableRegistry;
 use MeTools\TestSuite\TestCase;
-use MeTools\TestSuite\Traits\MockTrait;
+use PHPUnit\Framework\MockObject\MockObject;
 
 /**
  * MockTraitTest class
  */
 class MockTraitTest extends TestCase
 {
-    use MockTrait;
-
     /**
      * Tests for `getControllerAlias()` method
      * @test
@@ -44,7 +42,7 @@ class MockTraitTest extends TestCase
     public function testGetMockForComponent()
     {
         $Mock = $this->getMockForComponent('Cake\Controller\Component\FlashComponent', null);
-        $this->assertIsMock($Mock);
+        $this->assertInstanceOf(MockObject::class, $Mock);
     }
 
     /**
@@ -54,11 +52,11 @@ class MockTraitTest extends TestCase
     public function testGetMockForController()
     {
         $Mock = $this->getMockForController('App\Controller\PagesController', null);
-        $this->assertIsMock($Mock);
+        $this->assertInstanceOf(MockObject::class, $Mock);
         $this->assertEquals('Pages', $Mock->getName());
 
         $Mock = $this->getMockForController('App\Controller\PagesController', null, 'MyController');
-        $this->assertIsMock($Mock);
+        $this->assertInstanceOf(MockObject::class, $Mock);
         $this->assertEquals('MyController', $Mock->getName());
     }
 
@@ -70,25 +68,5 @@ class MockTraitTest extends TestCase
     public function testGetMockForControllerNoExistingClass()
     {
         $this->getMockForController('App\Controller\NoExistingController');
-    }
-
-    /**
-     * Tests for `getMockForTable()` method
-     * @test
-     */
-    public function testGetMockForTable()
-    {
-        $Mock = $this->getMockForTable('App\Model\Table\PagesTable', null);
-        $this->assertIsMock($Mock);
-        $this->assertEquals('Pages', $Mock->getAlias());
-        $this->assertEquals('Cake\ORM\Entity', $Mock->getEntityClass());
-        $this->assertNotEquals('App\Model\Entity\Page', $Mock->getEntityClass());
-        $this->assertTrue(TableRegistry::getTableLocator()->exists($Mock->getAlias()));
-
-        $Mock = $this->getMockForTable('App\Model\Table\PostsTable', null);
-        $this->assertIsMock($Mock);
-        $this->assertEquals('Posts', $Mock->getAlias());
-        $this->assertEquals('App\Model\Entity\Post', $Mock->getEntityClass());
-        $this->assertTrue(TableRegistry::getTableLocator()->exists($Mock->getAlias()));
     }
 }